+++ app/models/device-network-type-link/index.js
@@ -106,16 +106,31 @@ async function update (ctx, { where, data, origin }) {
     data = serializeNwkSettings(data)
   }
   const rec = await ctx.db.update({ where, data })
-  await ctx.$m.networkType.forAllNetworks({
-    networkTypeId: rec.networkType.id,
-    op: async network => ctx.$m.networkDeployment.updateByQuery({
-      where: { network: { id: network.id }, device: rec.device },
-      data: {
-        status: origin && origin.network.id === network.id ? 'SYNCED' : 'UPDATED'
-      }
+  if (origin) {
+    await ctx.$m.networkDeployment.updateByQuery({
+      where: { network: { id: origin.network.id }, device: { id: rec.id } },
+      data: { status: 'SYNCED' }
     })
+  }
+  await ctx.$self.push({
+    deviceNetworkTypeLink: rec,
+    omitNetworks: origin ? [origin.network.id] : []
   })
   return parseNwkSettings(rec)
+}
+
+async function update (ctx, { where, data, origin }) {
+  if (data.networkSettings) {
+    data = { ...data, networkSettings: prune(data.networkSettings) }
+    validateNwkSettings(data.networkSettings)
+  }
+  const rec = await ctx.db.update({ where, data })
+
+  await ctx.$self.push({
+    deviceProfile: rec,
+    omitNetworks: origin ? [origin.network.id] : []
+  })
+  return rec
 }
 
 async function upsert (ctx, { data, ...args }) {
@@ -144,17 +159,17 @@ async function remove (ctx, { id }) {
   await ctx.db.remove(id)
 }
 
-async function pushDeviceNetworkTypeLink (ctx, id) {
-  var rec = await ctx.db.load({ where: { id } })
-  rec.remoteAccessLogs = await ctx.$m.networkType.forAllNetworks({
+async function pushDeviceNetworkTypeLink (ctx, { id, deviceNetworkTypeLink: rec, omitNetworks = [] }) {
+  if (!rec) {
+    rec = await ctx.db.load({ where: { id } })
+  }
+  await ctx.$m.networkType.forAllNetworks({
     networkTypeId: rec.networkType.id,
-    op: network => ctx.$m.networkProtocol.pushDevice({
-      network,
-      deviceId: rec.device.id,
-      networkSettings: rec.networkSettings
-    })
+    op: async network => {
+      if (omitNetworks.includes(network.id)) return
+      return ctx.$m.networkProtocol.pushDevice({ network, deviceId: rec.id })
+    }
   })
-  return rec
 }
 
 async function findByDevEUI (ctx, { devEUI, networkTypeId }) {
@@ -188,7 +203,7 @@ module.exports = {
     upsert,
     remove,
     removeMany,
-    pushDeviceNetworkTypeLink,
+    push: pushDeviceNetworkTypeLink,
     findByDevEUI
   },
   fragments
